requirements-dev.txt
pytest>=9.0
pyyaml>=6.0
